Extend threading tests
Rework worker methods for threading tests to be able to handle varying block-of-element sizes.
Put threading test into seperate function and be able to call this with varying block-of-element sizes.
Add extra test for Poke(). Buffer has 3 states: empty, full, somewhat in between - the calculation for internal variable 'available' must handle this properly.