Add tests for TryPop()
Add tests for the TryPop() method. Remove the Peek() and Poke() method. Rationale: these are redundant. Size() can be requested as indication of number of elements. If TryPush() fails calling side must wait for consumer before there is room - no different from Poke(). If data does not fit most likely consumer is not fast enough with removal of data.